Guard controllers with JWT authentication scheme. Add more try-catch blocks. Change way of fetching user ID. Modify some of the returned status codes

+++ src/eduProjectWebAPI/Controllers/NotificationsController.cs
@@ -1,14 +1,18 @@
 ﻿using eduProjectWebAPI.Data;
+using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace eduProjectWebAPI.Controllers
 {
     [ApiController]
     [Route("notifications")]
+    [Authorize(AuthenticationSchemes = JwtBearerDefaults.AuthenticationScheme)]
     public class NotificationsController : ControllerBase
     {
         private readonly INotificationsRepository notifications;
@@ -21,29 +25,77 @@ namespace eduProjectWebAPI.Controllers
         [HttpGet("author/{authorId:int}/applications")]
         public async Task<ActionResult<ICollection<int>>> GetReceivedApplicationsNotification(int authorId)
         {
-            var result = await notifications.GetReceivedApplicationsNotification(authorId);
-            return result.ToList();
+            try
+            {
+                if (int.Parse(HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier)) == authorId)
+                {
+                    var result = await notifications.GetReceivedApplicationsNotification(authorId);
+                    return result.ToList();
+                }
+                else
+                    return Forbid();
+            }
+            catch (Exception e)
+            {
+                return BadRequest(e.Message + "\n" + e.StackTrace);
+            }
         }
 
         [HttpGet("user/{userId:int}/applications")]
         public async Task<ActionResult<ICollection<int>>> GetSentApplicationsNotification(int userId)
         {
-            var result = await notifications.GetSentApplicationsNotification(userId);
-            return result.ToList();
+            try
+            {
+                if (int.Parse(HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier)) == userId)
+                {
+                    var result = await notifications.GetSentApplicationsNotification(userId);
+                    return result.ToList();
+                }
+                else
+                    return Forbid();
+            }
+            catch (Exception e)
+            {
+                return BadRequest(e.Message + "\n" + e.StackTrace);
+            }
         }
 
         [HttpDelete("author/{authorId:int}/applications")]
         public async Task<ActionResult> DeleteReceivedApplicationsNotification(int authorId)
         {
-            await notifications.DeleteReceivedApplicationsNotification(authorId);
-            return NoContent();
+            try
+            {
+                if (int.Parse(HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier)) == authorId)
+                {
+                    await notifications.DeleteReceivedApplicationsNotification(authorId);
+                    return NoContent();
+                }
+                else
+                    return Forbid();
+            }
+            catch (Exception e)
+            {
+                return BadRequest(e.Message + "\n" + e.StackTrace);
+            }
         }
 
         [HttpDelete("user/{userId:int}/applications")]
         public async Task<ActionResult> DeleteSentApplicationsNotification(int userId)
         {
-            await notifications.DeleteSentApplicationsNotification(userId);
-            return NoContent();
+            try
+            {
+                if (int.Parse(HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier)) == userId)
+                {
+                    await notifications.DeleteSentApplicationsNotification(userId);
+                    return NoContent();
+                }
+                else
+                    return Forbid();
+            }
+            catch (Exception e)
+            {
+                return BadRequest(e.Message + "\n" + e.StackTrace);
+            }
         }
     }
 }
